copywriter/requirements.txt
reflex==0.8.25
openai>=1.0.0
python-dotenv>=1.0.0